// src/app/models/game-stats.ts
export class GameStats {
    shots: number = 0;
    goals: number = 0;
    points: number;
    powerPlays: number = 0;
    penaltyMinutes: number = 0;
    plusMinus: number = 0;
    hits: number = 0;
    blocks: number = 0;
    faceoffs: number = 0;
    timeOnAttack: number = 0;
    timeOnIce: number = 0;
    saves: number = 0;
    goalsAllowed: number = 0;

    constructor() {

    }
}